requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
deep-translator==1.11.4
dnspython==2.4.2
exceptiongroup==1.1.3
frozenlist==1.4.0
geographiclib==2.0
geopy==2.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hstspreload==2023.1.1
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==2.10
libretranslatepy==2.1.1
lxml==4.9.3
magic-filter==1.0.11
motor==3.3.1
multidict==6.0.4
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.5
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
